requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.0
Flask-Talisman==1.1.0
SQLAlchemy==2.0.40
Werkzeug==3.1.3
python-dotenv==1.0.0
email-validator==2.1.0.post1
gunicorn==21.2.0
pytest==7.4.3
